Run doxy, add new sound for level 3, fixing appear level complete, fixing score and hit test

+++ GameLib/DeclarationMeter.cpp
@@ -17,7 +17,6 @@ const double MaxNeedleRotation = 0.9;
 /// as a percentage of the height of the image.
 const double NeedlePivotYOffset = 0.80;
 
-
 using namespace std;
 
 /// All images directory
@@ -27,7 +26,7 @@ wstring const ImageDir = L"images/";
  * Constructor
  * @param game Game this declaration is a member of
  */
-DeclarationMeter::DeclarationMeter(Game *game): Declaration(game)
+DeclarationMeter::DeclarationMeter(Game *game) : Declaration(game)
 {
 }
 
@@ -37,50 +36,51 @@ DeclarationMeter::DeclarationMeter(Game *game): Declaration(game)
  */
 void DeclarationMeter::XmlLoad(wxXmlNode *node)
 {
-    wxString coverFilename = ImageDir + node->GetAttribute(L"cover","");
+    wxString coverFilename = ImageDir + node->GetAttribute(L"cover", "");
     mCoverBitmap = make_unique<wxBitmap>(
         coverFilename, wxBITMAP_TYPE_ANY);
-    wxString needleFilename = ImageDir + node->GetAttribute(L"needle","");
+    wxString needleFilename = ImageDir + node->GetAttribute(L"needle", "");
     mNeedleBitmap = make_unique<wxBitmap>(
         needleFilename, wxBITMAP_TYPE_ANY);
 
     Declaration::XmlLoad(node);
 }
 
-void DeclarationMeter:: Draw(std::shared_ptr <wxGraphicsContext> gp, double x, double y, bool before)
+void DeclarationMeter::Draw(std::shared_ptr<wxGraphicsContext> gp, double x, double y, bool before)
 {
     Declaration::Draw(gp, x, y, before);
 
-    int wid = this->GetWidth();
-    int hit = this->GetHeight();
+    double wid = this->GetWidth();
+    double hit = this->GetHeight();
     int needlePivotY = (int)(hit * NeedlePivotYOffset);
 
     gp->PushState();
-    gp->Translate(x, y + needlePivotY - hit/2);
+    gp->Translate(x, y + needlePivotY - hit / 2);
     double playedNote = this->GetGame()->GetPlayedNote();
     double totalNote = this->GetGame()->GetTotalNote();
 
-    if (totalNote == 0){
+    if(totalNote == 0)
+    {
         gp->Rotate(MaxNeedleRotation);
         gp->DrawBitmap(*mNeedleBitmap,
-                       -wid/2,
+                       -wid / 2,
                        -needlePivotY,
                        wid, hit);
     }
-    else{
-        gp->Rotate( min(-MaxNeedleRotation + playedNote/totalNote * 2 * MaxNeedleRotation,MaxNeedleRotation));
+    else
+    {
+        gp->Rotate(min(-MaxNeedleRotation + playedNote / totalNote * 2 * MaxNeedleRotation, MaxNeedleRotation));
         gp->DrawBitmap(*mNeedleBitmap,
-                       -wid/2,
+                       -wid / 2,
                        -needlePivotY,
                        wid, hit);
     }
-
 
     gp->PopState();
 
     gp->DrawBitmap(*mCoverBitmap,
-                   int(x -this->GetWidth() / 2),
-                   int(y -this->GetHeight() / 2),
+                   int(x - this->GetWidth() / 2),
+                   int(y - this->GetHeight() / 2),
                    this->GetWidth(), this->GetHeight());
 }
